pyproject.toml
[project]
name = "wellread"
version = "1.0.0"
description = "RSS feed monitoring Slackbot with AI-powered summarization"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
keywords = ["rss", "slack", "bot", "ai", "summarization"]
authors = [
    { name = "WellRead Bot" }
]

dependencies = [
    "anthropic>=0.40.0",
    "slack-sdk>=3.27.0",
    "feedparser>=6.0.11",
    "python-dateutil>=2.9.0",
    "openai>=1.0.0",
    "numpy>=1.24.0",
    "tqdm>=4.65.0",
]

[tool.uv]
dev-dependencies = []

[tool.uv.sources]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]
